B7abiertosREST/requirements.txt
fastapi>=0.85.0
uvicorn>=0.18.3
python-dotenv>=0.21.0
requests>=2.28.1
autopep8>=1.7.0
pandas>=1.5.0